+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/hibernate/common/CdmEntityDaoBase.java
@@ -839,7 +839,7 @@ public abstract class CdmEntityDaoBase<T extends CdmBase> extends DaoBase implem
     }
 
 
-    protected int countByParam(Class<? extends T> clazz, String param, String queryString, MatchMode matchmode, List<Criterion> criterion) {
+    protected long countByParam(Class<? extends T> clazz, String param, String queryString, MatchMode matchmode, List<Criterion> criterion) {
         Criteria criteria = null;
 
         if(clazz == null) {
@@ -866,8 +866,7 @@ public abstract class CdmEntityDaoBase<T extends CdmBase> extends DaoBase implem
 
         criteria.setProjection(Projections.rowCount());
 
-//        List<T> result = criteria.list();
-        return ((Number)criteria.uniqueResult()).intValue();
+        return ((Number)criteria.uniqueResult()).longValue();
     }
