+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/molecular/IAmplificationDao.java
@@ -43,7 +43,7 @@ public interface IAmplificationDao extends IAnnotatableDao<Amplification>{
      * @param criteria extra restrictions to apply
      * @return a count of instances of type T matching the queryString
      */
-    public Integer countByTitle(String queryString, MatchMode matchmode, List<Criterion> criteria);
+    public long countByTitle(String queryString, MatchMode matchmode, List<Criterion> criteria);
 
     /**
     * Return a List of {@link Amplification}s matching the given query string, optionally with a particular MatchMode
